+++ src/Bartlett/CompatInfo/Sniffs/FunctionDeclarations/ParamTypeDeclarationSniff.php
@@ -79,7 +79,9 @@ final class ParamTypeDeclarationSniff extends SniffAbstract
             // type hint object required at least PHP 5.0
             $this->updateVersion('5.0.0', $versions['php.min']);
 
-            if ($param->type instanceof Node\NullableType) {
+            if ($param->type instanceof Node\UnionType) {
+                $this->updateVersion('8.0.0', $versions['php.min']);
+            }  elseif ($param->type instanceof Node\NullableType) {
                 // @link https://www.php.net/manual/en/migration71.new-features.php#migration71.new-features.nullable-types
                 $this->updateVersion('7.1.0', $versions['php.min']);
             } else {